moved components into folder, changed imports
