+++ src/Pages/HomePage.tsx
@@ -1,16 +1,21 @@
-import React from 'react';
-import Footer from '../Components/footer'; // Adjust the path as needed based on your project structure
-import FindUsComponent from '../Components/FindUsComponent';
+import React from "react";
+import FindUsComponent from "../Components/FindUsComponent";
+import HeaderSection from "../Components/Header";
+import Footer from "../Components/footer"; // Adjust the path as needed based on your project structure
 
 const HomePage: React.FC = () => {
-    return (
-      <>
-        <header></header>
-        <main><FindUsComponent></FindUsComponent></main>
-        <footer>
-            <Footer />
-        </footer>
-      </>
+  return (
+    <>
+      <header>
+        <HeaderSection />
+      </header>
+      <main>
+        <FindUsComponent></FindUsComponent>
+      </main>
+      <footer>
+        <Footer />
+      </footer>
+    </>
   );
 };
